Allow ability to update root index page under gh-pages
Currently, we only support copying of docs to either latest, or snapshot, or individual version folders. But, there is no way to update the root index.html page under gh-pages branch. This commit will follow convention that if there is a file named "ghpages.html" under build/docs folder then it will copy to the root with name index.html.